+++ authority-portal-frontend/src/app/pages/participant-register-own-connector-page/participant-register-own-connector-page/participant-register-own-connector-page.component.ts
@@ -3,7 +3,9 @@ import {FormBuilder, FormGroup, Validators} from '@angular/forms';
 import {Subject, takeUntil} from 'rxjs';
 import {Store} from '@ngxs/store';
 import {CreateConnectorRequest} from '@sovity.de/authority-portal-client';
+import * as forge from 'node-forge';
 import {APP_CONFIG, AppConfig} from 'src/app/core/config/app-config';
+import {CertificateGenerateService} from 'src/app/shared/services/certificate-generate.service';
 import {
   Reset,
   Submit,
@@ -29,11 +31,17 @@ export class ParticipantRegisterOwnConnectorPageComponent
 {
   state = DEFAULT_PARTICIPANT_REGISTER_OWN_CONNECTOR_STATE;
   group = this.buildFormGroup();
+  keyPair!: forge.pki.rsa.KeyPair;
+  publicKeyPem!: string;
+  privateKeyPem!: string;
+  isDownloadActive: boolean = false;
+  hasDownloadedKey: boolean = false;
 
   constructor(
     @Inject(APP_CONFIG) public config: AppConfig,
     private store: Store,
     private formBuilder: FormBuilder,
+    private certificateGenerateService: CertificateGenerateService,
   ) {}
 
   ngOnInit(): void {
@@ -66,7 +74,37 @@ export class ParticipantRegisterOwnConnectorPageComponent
     return this.group.value as ParticipantRegisterOwnConnectorPageFormValue;
   }
 
+  generateCertificate() {
+    this.keyPair = this.certificateGenerateService.generateKeyPair(2048);
+    this.publicKeyPem = this.certificateGenerateService.publicKeyToPem(
+      this.keyPair.publicKey,
+    );
+    this.privateKeyPem = this.certificateGenerateService.privateKeyToPem(
+      this.keyPair.privateKey,
+    );
+    this.group.controls['certificate'].setValue(this.publicKeyPem);
+    this.isDownloadActive = true;
+  }
+
+  downloadPrivateKey() {
+    this.hasDownloadedKey = true;
+    const blob = new Blob([this.privateKeyPem], {type: 'text/plain'});
+    const url = window.URL.createObjectURL(blob);
+
+    const a = document.createElement('a');
+    a.href = url;
+    a.download = 'private_key.pem';
+
+    // Trigger the download
+    a.click();
+
+    window.URL.revokeObjectURL(url);
+  }
+
   submit(): void {
+    if (!this.hasDownloadedKey) {
+      this.downloadPrivateKey();
+    }
     const formValue: ParticipantRegisterOwnConnectorPageFormValue = this.value;
     const request: CreateConnectorRequest = {
       name: formValue.name!,